\name{dkDNA-package}
\alias{dkDNA-package}
\alias{dkDNA}
\docType{package}
\title{
Diffusion kernels on a set of genotypes
}

\description{
Compute diffusion kernels on DNA polymorphisms, including SNP and binary genotypes. 
}

\author{
Gota Morota \email{morota@ansci.wisc.edu} and Masanori Koyama \email{koyama@math.wisc.edu}
}

\references{
  Kondor R and Lafferty J: (2002) Diffusion Kernels on Graphs and Other
  Discrete Input Spaces. ICML. 

  Morota G and Koyama M: (2012) Diffusion kernels on SNP data embedded in a
  non-Euclidean metric. In preparation. 
}

\keyword{ package }

