\name{dkDNA-package}
\alias{dkDNA-package}
\alias{dkDNA}
\docType{package}
\title{
Diffusion kernels on a set of genotypes
}

\description{
Compute diffusion kernels on DNA polymorphisms, including SNP and binary genotypes. 
}

\author{
Gota Morota \email{morota@unl.edu} and Masanori Koyama \email{koyama.masanori@gmail.com}
}

\references{
  Kondor R and Lafferty J: (2002) Diffusion Kernels on Graphs and Other
  Discrete Input Spaces. ICML. 

  Morota G, Koyama M, Rosa GJM, Weigel KA, and Gianola D. (2013). Predicting complex traits using a diffusion kernel on genetic markers with an application to dairy cattle and wheat data. Genetics Selection Evolution. 45:17. 

}

\keyword{ package }

